+++ addons/ammunition/ammo/357.hpp
@@ -1,0 +1,3 @@
+class CLASS(357_Magnum): B_45ACP_Ball {
+    hit = 8;
+};
