+++ addons/ammunition/ammo/44.hpp
@@ -1,0 +1,3 @@
+class CLASS(44_Magnum): CLASS(357_Magnum) {
+    hit = 11;
+};
