+++ addons/ammunition/ammo/545x39.hpp
@@ -1,0 +1,10 @@
+class CLASS(545x39_BP): B_545x39_Ball_F {
+    hit = 9;
+    caliber = 1.19;
+};
+
+class CLASS(545x39_BT): B_545x39_Ball_F {
+    model = "\A3\Weapons_f\Data\bullettracer\tracer_red";
+    hit = 9.4;
+    caliber = 1.22;
+};
