+++ addons/ammunition/ammo/57x28.hpp
@@ -1,0 +1,4 @@
+class CLASS(57x28_SS190): B_570x28_Ball {
+    hit = 9.4;
+    caliber = 0.89;
+};
