+++ addons/ammunition/ammo/58x42.hpp
@@ -1,0 +1,4 @@
+class CLASS(580x42_DBP88): B_580x42_Ball_F {
+    hit = 9.6;
+    caliber = 1.1;
+};
