+++ addons/ammunition/ammo/762x39.hpp
@@ -1,0 +1,5 @@
+class CLASS(762x39_BP): B_762x39_Ball_F {
+    hit = 14;
+    caliber = 2.1;
+    model = "\A3\Weapons_f\Data\bullettracer\tracer_yellow";
+};
