+++ addons/ammunition/ammo/762x54r.hpp
@@ -1,0 +1,5 @@
+class CLASS(762x54_AP): CUP_B_762x54_Ball_White_Tracer {
+    hit = 14.6;
+    caliber = 2.4;
+    model = "\A3\Weapons_f\Data\bullettracer\tracer_yellow";
+};
