+++ addons/ammunition/ammo/9x19.hpp
@@ -1,0 +1,4 @@
+class CLASS(9x19_7N21): B_9x21_Ball_Tracer_Green {
+    caliber = 1.4;
+    hit = 6.5;
+};
