+++ addons/ammunition/magazines/12g.hpp
