+++ addons/ammunition/magazines/545x39.hpp
