+++ addons/ammunition/magazines/556x45.hpp
