+++ addons/ammunition/magazines/580x42.hpp
