+++ addons/ammunition/magazines/762x39.hpp
