+++ addons/ammunition/magazines/762x51.hpp
