+++ addons/ammunition/magazines/762x54r.hpp
